# src/openapi/mobile-v1.yaml
openapi: 3.0.3
info:
  title: CRMtree API — mobile contract
  version: 1.0.0
  description: |
    The endpoints the CRMtree mobile app (ADR 001) is allowed to depend on.
    The Dart client in `crmtree-mobile` is generated from this file.

    Rules:
    - Old app builds stay installed for weeks. A change that removes or renames
      a field, tightens a type, or adds a required request field is BREAKING
      and needs a new endpoint or version (CI runs `oasdiff` against master).
    - Response objects are open (`additionalProperties: true`): the web API
      returns many more fields than listed. Only the fields documented here
      are part of the contract; the app must ignore the rest.
    - Money values (`value_pln`, `contract_value`) are decimal STRINGS, as
      PostgreSQL NUMERIC serializes — never parse them as binary floats.
    - Every session talks to its own tenant host (`{slug}.crmtree.pl/api`).
      `POST /auth/mobile/login` and `GET /public/app-config` are the only
      calls made on the tenant-less host.
servers:
  - url: https://{tenant}.crmtree.pl/api
    description: Production, per-tenant host
    variables:
      tenant:
        default: app
        description: Tenant slug; `app` for the tenant-less host used before login
  - url: https://int.crmtree.pl/api
    description: Integration environment (tenant comes from the token)
...
security:
  - bearerAuth: []
paths:
...
  /crm/leads:
    get:
      tags: [leads]
      operationId: listLeads
      summary: Leads visible to the user (own leads for a salesperson, the team for a manager)
      parameters:
        - { name: stage, in: query, schema: { $ref: '#/components/schemas/LeadStage' }, description: Without it archived leads are hidden }
        - { name: mine, in: query, schema: { type: boolean } }
        - { name: hot, in: query, schema: { type: boolean } }
        - { name: hold_only, in: query, schema: { type: boolean } }
        - { name: source, in: query, schema: { type: string }, description: 'A source value, or several separated by commas' }
        - { name: search, in: query, schema: { type: string }, description: 'Matches company, contact, email, NIP, phone, notes and tags' }
        - { name: assigned_to, in: query, schema: { type: string }, description: User id or comma-separated ids (managers only) }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 5000, default: 50 } }
      responses:
        '200':
          description: A page of leads, most recently updated first
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeadListResponse' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [leads]
      operationId: createLead
      summary: Add a lead
      description: |
        A salesperson always becomes the owner (`assigned_to` is ignored).
        A NIP already used by another lead or partner returns 409.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadCreate' }
      responses:
        '201':
          description: Created lead
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lead' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /admin/settings:
    get:
      tags: [app]
      operationId: getSettings
      summary: The company's settings as a flat key-value map
      description: |
        The app reads `crm_lost_reasons` from it: a JSON-encoded array of the
        reasons offered when a lead is marked as lost.
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettingsResponse' }
        '401': { $ref: '#/components/responses/Error' }
...
  /crm/partners:
    get:
      tags: [partners]
      operationId: listPartners
      summary: Partners (customers)
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string }, description: 'active, inactive or churned; a partner without a CRM status counts as active' }
        - { name: manager_id, in: query, schema: { type: string }, description: Managers only }
        - { name: group_name, in: query, schema: { type: string } }
        - { name: industry, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, default: 50 } }
        - { name: sort, in: query, schema: { type: string, enum: [company, status, contract_signed, contract_value, created_at, industry, group_name, manager_name, contract_expires], default: company } }
        - { name: dir, in: query, schema: { type: string, enum: [asc, desc], default: asc } }
      responses:
        '200':
          description: A page of partners
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PartnerListResponse' }
        '401': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: true
      properties:
        error: { type: string, description: 'Human-readable message (Polish), safe to show' }
        code: { type: string, description: 'Machine-readable code when the app must react, e.g. TOKEN_EXPIRED' }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    LeadStage:
      type: string
      enum: [new, qualification, presentation, offer, negotiation, closed_won, closed_lost, archived]
    Lead:
      type: object
      required: [id, company, stage, hot, created_at, updated_at]
      additionalProperties: true
      properties:
        id: { type: integer }
        company: { type: string }
        contact_name: { type: string, nullable: true }
        contact_title: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        website: { type: string, nullable: true }
        logo_url: { type: string, nullable: true, description: 'Set when the lead has a logo; fetch it from /crm/leads/{id}/logo-img' }
        nip: { type: string, nullable: true }
        industry: { type: string, nullable: true }
        source: { type: string, nullable: true }
        stage: { $ref: '#/components/schemas/LeadStage' }
        value_pln: { type: string, nullable: true, description: Decimal string }
        probability: { type: integer, nullable: true, minimum: 0, maximum: 100 }
        close_date: { type: string, nullable: true, description: ISO timestamp of the planned close date }
        assigned_to: { type: string, format: uuid, nullable: true }
        tags:
          type: array
          nullable: true
          items: { type: string }
        notes: { type: string, nullable: true }
        hot: { type: boolean }
        lost_reason: { type: string, nullable: true }
        hold_active: { type: boolean }
        hold_until: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    LeadSummary:
      allOf:
        - $ref: '#/components/schemas/Lead'
        - type: object
          additionalProperties: true
          properties:
            assigned_to_name: { type: string, nullable: true }
            activity_count: { type: integer }
            non_email_activity_count: { type: integer, description: Open non-email activities }
            next_activity_type: { type: string, nullable: true, description: 'meeting, call or task: the earliest open dated one, overdue first' }
            next_activity_title: { type: string, nullable: true }
            next_activity_at: { type: string, format: date-time, nullable: true }
            new_email_count: { type: integer }
            unread_sms_count: { type: integer }
            unread_whatsapp_count: { type: integer }
            missed_call_count: { type: integer }
    LeadListResponse:
      type: object
      required: [data, total, page, limit, pages]
      additionalProperties: true
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/LeadSummary' }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }
...
    LeadCreate:
      type: object
      required: [company]
      properties:
        company: { type: string, minLength: 1 }
        contact_name: { type: string }
        contact_title: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string }
        nip: { type: string, nullable: true }
        value_pln: { type: number, nullable: true, minimum: 0 }
        notes: { type: string }
        hot: { type: boolean }
    SettingsResponse:
      type: object
      required: [settings]
      additionalProperties: true
      properties:
        settings:
          type: object
          additionalProperties: true
          properties:
            crm_lost_reasons: { type: string, description: JSON-encoded array of strings }
...
    PartnerSummary:
      type: object
      required: [crm_uuid, company]
      additionalProperties: true
      properties:
        crm_uuid: { type: string, format: uuid, description: Use this as partnerId in paths }
        company: { type: string }
        status: { type: string, nullable: true }
        manager_id: { type: string, format: uuid, nullable: true }
        manager_name: { type: string, nullable: true }
        contract_value: { type: string, nullable: true, description: Decimal string }
        tags:
          type: array
          nullable: true
          items: { type: string }
        non_email_activity_count: { type: integer }
        new_email_count: { type: integer }
        unread_sms_count: { type: integer }
        unread_whatsapp_count: { type: integer }
        missed_call_count: { type: integer }
        churn_risk: { type: string, nullable: true }
        churn_score: { type: number, nullable: true }
        dwh_partner_id: { type: integer, nullable: true, description: Key of the partner in the sales report }
        active_since: { type: string, format: date-time, nullable: true, description: 'When the customer started: from the data warehouse, otherwise when the partner was created in the CRM' }
        website: { type: string, nullable: true }
        logo_url: { type: string, nullable: true, description: 'Set when the partner has a logo; fetch it from /crm/partners/{partnerId}/logo-img' }
        next_activity_type: { type: string, nullable: true, description: 'meeting, call or task: the earliest open dated one, overdue first' }
        next_activity_title: { type: string, nullable: true }
        next_activity_at: { type: string, format: date-time, nullable: true }
...
    PartnerListResponse:
      type: object
      required: [data, total, page, limit]
      additionalProperties: true
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/PartnerSummary' }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...